i forgot to fix this

+++ SonicMania/GameLink.h
@@ -1647,7 +1647,7 @@ typedef struct {
     int32 (*PlaySfx)(uint16 sfx, int32 loopPoint, int32 priority);
     void (*StopSfx)(uint16 sfx);
     int32 (*PlayStream)(const char *filename, uint32 channel, uint32 startPos, uint32 loopPoint, bool32 loadASync);
-    void (*SetChannelAttributes)(uint8 channel, float volume, float pan, float speed);
+    void (*SetChannelAttributes)(uint32 channel, float volume, float pan, float speed);
     void (*StopChannel)(uint32 channel);
     void (*PauseChannel)(uint32 channel);
     void (*ResumeChannel)(uint32 channel);
